implemented draggable blocks
