+++ package.json
@@ -28,6 +28,7 @@
     "node-sass": "^4.14.1",
     "react": "^16.14.0",
     "react-dnd": "^11.1.3",
+    "react-dnd-html5-backend": "^11.1.3",
     "react-dom": "^16.14.0",
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
