+++ src/utils/icons/Delete.js
@@ -8,7 +8,7 @@ const Delete = () => (
     fill="none"
     xmlns="http://www.w3.org/2000/svg"
   >
-    <g clip-path="url(#clip18)">
+    <g clipPath="url(#clip18)">
       <path
         d="M10.8449 6.15912C10.625 6.15912 10.4468 6.33734 10.4468 6.55725V14.0818C10.4468 14.3016 10.625 14.48 10.8449 14.48C11.0648 14.48 11.243 14.3016 11.243 14.0818V6.55725C11.243 6.33734 11.0648 6.15912 10.8449 6.15912Z"
         fill="#F9F9F9"
