+++ src/views/home/components/Choice.js
@@ -3,7 +3,7 @@ import { connect } from "react-redux";
 
 import { changeSelection } from "../../../actions/options.action";
 import { OPTIONS } from "../data";
-import Block from "./blocks/Block";
+import Blocks from "./blocks/Blocks";
 import Options from "./options/Options";
 import Structure from "./structure/Structure";
 import Settings from './settings/Settings';
@@ -20,7 +20,7 @@ class Choice extends Component {
             selection={this.props.selection}
             onHandleSelection={this.onHandleSelection}
           />
-          {this.props.selection === OPTIONS[0] && <Block />}
+          {this.props.selection === OPTIONS[0] && <Blocks />}
           {this.props.selection === OPTIONS[1] && <Structure />}
           {this.props.selection === OPTIONS[2] && <Settings />}
         </div>
