+++ src/views/home/components/blocks/Block.js
@@ -1,27 +1,24 @@
 import React from "react";
+import { useDrag } from "react-dnd";
 
-import Cards from "../../../../components/molecules/Cards";
-import { CARDS } from "../../data";
+import Card from "../../../../components/molecules/Cards";
+import { ITEMS } from "../../data";
 
-const Block = () => {
-  const generateBlocks = () => {
-    const rows = [...Array(Math.ceil(CARDS.length / 3))];
-    const blockRows = rows.map((row, index) =>
-      CARDS.slice(index * 3, index * 3 + 3)
-    );
-    const blocks = blockRows.map((row, index) => (
-      <div className="a-row" key={index}>
-        {row.map((block, idx) => (
-          <div className="a-col-1-of-3" key={`${index}-${idx}`}>
-            <Cards {...block} />
-          </div>
-        ))}
-      </div>
-    ));
-    return blocks;
-  };
-
-  return <div className="u-padding-custom-1">{generateBlocks()}</div>;
+const Block = ({ block }) => {
+  const [{ opacity }, dragRef] = useDrag({
+    item: {
+      type: ITEMS.BLOCK,
+      content: block,
+    },
+    collect: (monitor) => ({
+      opacity: !!monitor.isDragging() ? "0.5" : "1",
+    }),
+  });
+  return (
+    <span ref={dragRef} style={{ opacity }} className="card__blockWrapper">
+      <Card {...block} />
+    </span>
+  );
 };
 
 export default Block;
